feat: add NoteEntity, Converters and NoteXmlBuilder
Add the Note Entity and assign it to the Room Database. Add a Date Converter to allow Room to handle the type. Add a NoteItemConverter that allows for the class to be turned into a String representation ('xml'). Add a NoteItemXmlBuilder that can be used to change the NoteItems to xml to be written to the database and read/ interpreted.